Created a Wrapper class for Graph called TileSymbolGraph. It is based off of SymbolGraph from algs4 but uses Tiles as vertexes instead of Strings.
Also updated Tile class to implement comparable. Tiles are compared by their x values first, and y values second in case the x values are the same.

Created a File called tileConnections.txt that contains the attributes of every needed tile and the connections each of them need in the graph.