Added Icons to all Piece Classes. All Files are in the Resources folder.
Made small changes to fields across several classes to better suit their purpose.

In Board.java, Created and added pieces to their default positions.

In ChessApp.java, Create the main Grid UI with the tiles, Buttons,actionListeners, Icons, etc... The Chess functionality is not there yet, but this is close to the final look of the program.

In TileSymbolGraph.java, Added Ability to get a Tile object based on its name.

Added main method in Board.java for testing purposes. It will be removed later.